backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0

# Environment
python-dotenv>=1.0

# Database (for production)
psycopg2-binary>=2.9
dj-database-url>=2.0

# Authentication
social-auth-app-django>=5.4
requests>=2.31

# File uploads
Pillow>=10.0

# Production
gunicorn>=21.0
whitenoise>=6.6

# Type checking
mypy>=1.8
django-stubs>=4.2
djangorestframework-stubs>=3.14
types-requests>=2.31

# Development
# django-debug-toolbar>=4.2

# Testing
pytest>=7.4
pytest-django>=4.5
